Add stylesheet, button ui
In this patch of the rock paper scissors game,
we added some divs to the index file to give our
buttons a nice lil home, and also added a style
sheet to prevent the body from collapsing in on itself.
The buttons now run through the game, and you get a console
log result.